+++ infra/app.py
@@ -1,5 +1,5 @@
 import aws_cdk as cdk
-from api_stack import DomainsRegistrationStack
+from infra_stack import DomainsRegistrationStack
 
 app = cdk.App()
 DomainsRegistrationStack(app, "domains-registration-stack")
